// src/routes/routeName.d.ts
export declare const routes: {
    home: string;
    about: string;
    login: string;
    forgot_password: string;
    unauthorized: string;
    verify_otp: {
        student: string;
        parent: string;
        tutor: string;
        forgot_password: string;
    };
    register: {
        general: string;
        student: string;
        tutor: string;
        parent: string;
    };
    tutor: {
        list: string;
        detail: string;
        dashboard: string;
        schedule: string;
        class: {
            list: string;
            create: string;
            detail: string;
        };
        information: string;
        change_password: string;
        wallet: string;
        booking: {
            list: string;
            detail: string;
        };
        request: {
            list: string;
            detail: string;
        };
    };
    course: {
        list: string;
        detail: string;
    };
    student: {
        home: string;
        tutor: {
            list: string;
            detail: string;
            book: string;
        };
        course: {
            list: string;
            detail: string;
        };
        information: string;
    };
    parent: {
        home: string;
        tutor: {
            list: string;
            detail: string;
            book: string;
        };
        course: {
            list: string;
            detail: string;
        };
        information: string;
    };
    admin: {
        dashboard: string;
        student: {
            list: string;
            detail: string;
        };
        parent: {
            list: string;
            detail: string;
        };
        tutor: {
            list: string;
            detail: string;
        };
        class: {
            list: string;
        };
        notification: {
            list: string;
        };
        report: {
            list: string;
        };
        transaction: {
            tutor: {
                list: string;
            };
            student: {
                list: string;
            };
            parent: {
                list: string;
            };
        };
    };
};