// src/utils/fakeData.d.ts
export declare const fakeDataTutors: {
    id: number;
    name: string;
    specialty: string[];
    rating: number;
    totalReviews: number;
    profileImage: string;
}[];
export declare const fakeDataCourses: {
    id: number;
    classLevel: string;
    instructor: string;
    subject: string;
    format: string;
    address: string;
    startDate: Date;
    price: number;
    students: number;
    rating: number;
    reviews: number;
    image: string;
}[];